So in order to make sure you can't chop logs on someone else's claim, I had to move TreeFeller to clans plugin
