+++ progression/src/main/java/me/mykindos/betterpvp/progression/profession/skill/woodcutting/TreeFeller.java
@@ -2,29 +2,15 @@ package me.mykindos.betterpvp.progression.profession.skill.woodcutting;
 
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
-import me.mykindos.betterpvp.core.listener.BPvPListener;
 import me.mykindos.betterpvp.progression.Progression;
-import me.mykindos.betterpvp.progression.profession.woodcutting.WoodcuttingHandler;
-import me.mykindos.betterpvp.progression.profession.woodcutting.event.PlayerChopLogEvent;
-import me.mykindos.betterpvp.progression.profile.ProfessionProfileManager;
 import org.bukkit.Material;
-import org.bukkit.block.Block;
-import org.bukkit.entity.Player;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.Listener;
 
 @Singleton
-@BPvPListener
-public class TreeFeller extends WoodcuttingProgressionSkill implements Listener {
-
-    private final ProfessionProfileManager professionProfileManager;
-    private final WoodcuttingHandler woodcuttingHandler;
+public class TreeFeller extends WoodcuttingProgressionSkill {
 
     @Inject
-    public TreeFeller(Progression progression, ProfessionProfileManager professionProfileManager, WoodcuttingHandler woodcuttingHandler) {
+    public TreeFeller(Progression progression) {
         super(progression);
-        this.professionProfileManager = professionProfileManager;
-        this.woodcuttingHandler = woodcuttingHandler;
     }
 
     @Override
@@ -42,37 +28,5 @@ public class TreeFeller extends WoodcuttingProgressionSkill implements Listener 
     @Override
     public Material getIcon() {
         return Material.GOLDEN_AXE;
-    }
-
-
-    @EventHandler
-    public void onPlayerChopsLog(PlayerChopLogEvent event) {
-        if (event.isCancelled()) return;
-
-        Player player = event.getPlayer();
-        professionProfileManager.getObject(player.getUniqueId().toString()).ifPresent(profile -> {
-            int skillLevel = getPlayerSkillLevel(profile);
-            if (skillLevel <= 0) return;
-
-            event.setCancelled(true);
-            fellTree(event.getChoppedLogBlock(), event, true);
-        });
-    }
-
-    public void fellTree(Block block, PlayerChopLogEvent event, boolean initialBlock) {
-        if (!initialBlock && woodcuttingHandler.didPlayerPlaceBlock(block)) return;
-
-        block.breakNaturally();
-        event.setAmountChopped(event.getAmountChopped() + 1);
-
-        for (int x = -1; x <= 1; x++) {
-            for (int z = -1; z <= 1; z++) {
-                Block targetBlock = block.getRelative(x, 1, z);
-
-                if(targetBlock.getType().name().contains("_LOG")) {
-                    fellTree(targetBlock, event, false);
-                }
-            }
-        }
     }
 }
